Annotations are being added to the map correctly.
I’m in the process of re-designing how the annotations work with the
Pin class and how they are then added to the map. That’s why this
commit looks very much like the previous commit. I made a bunch of
changes, only to come back around to where I was.