Removing pins is working.
There’s a discrepancy between the pin number when it is dropped and the
pin number when it is removed. This is because the Pin class’s total
number of pins is incremented and decremented each time a pin is added
or removed. However, pin numbers get reused, leading to situations
where two (or more) pins can share numbers. This doesn’t seem to be
affecting the ability to remove the correct pin, but it isn’t ideal.