Modified the "no images" label.
The text was changed and the label was moved further down in the view.
Made sure it looks good in iPhone 5s/6. It is also now visible by
default and being hidden when the view loads. Then, I’ll make it
visible if no images are found for a location.