Yep, you can go back and forth from the map and get photos for different locations.
It appears as though the collection view isn’t using 30 different
photos, though. And the tasks might not actually start until a cell
scrolls onscreen. The first 11 that appear with their activity
indicators offset might not ever get a photo returned to them, and the
other 19 might not ever start until one of them comes onscreen (and
then the first 11 will start, as well).