Commit before making a big change for a fix.
I’m having trouble figuring out how to make the collection view have a
dynamic number of cells based on the number of images found for the
location from Flickr. As a result, the current problem I’m experiencing
is that because the collection view is hardcoded to have a set number
of cells, if the result set is less than that number,
cellForItemAtIndexPath: will go out of bounds. However, as a temporary
fix, I can maybe have a default kind of cell get returned for the cells
that won’t house an image. I’m gonna commit before trying to implement
this, though.
Oh, by the way, I fixed the no results bug. And, a nice alert is
appearing when it happens.