Fixed a few things.
- Made the number of cells in the collection view dynamic.
- Made the cells update their images instantly.
- Handles a result set of one image correctly.

But there’s a new, weird bug: in a certain case, in the one instance
where I saw it, it was on a “middle” row of cells (from the top down;
ie., row four in a seven-row collection view), the middle row’s images
shifted between the cells in the row when it scrolled offscreen.
However, none of the other rows experienced this.