Should have merged this branch into Master long ago.
I got off on a lot of tangents, in the meantime. I’m currently stuck
trying to get my collection view to show thirty images instead of only
a number equal to the number of cells initially visible when the view
controller loads. I’m gonna try going back to the old way of doing it,
where I didn’t use two separate tasks, because it worked when I did it
that way. I’m gonna try collecting all thirty images and then returning
those in the completion handler, not the photo info. But I’m gonna
create a new branch to try it.