Still can't get it to work right.
The tasks for individual images are always running after returning the
array of images back to the photo album view controller. So, the array
is always empty and the collection view never has anything in it.
One last weird approach I can take it to return an array of tasks,
instead. Then in cellForItemAtIndexPath(), I can assign the task and
run it.