It ... works. It actually works 😁
This approach looks like it works. I’m saving the image data and then
creating new images with the image data when required by the collection
view. Scrolling the collection view confirms that cells are creating
new images when necessary or reusing an already created image. And
there are no image reuse bugs. And it only requires one Flickr
request!!!