Refactored requesting a new collection.
It appears to be working for result sets of max images (250). However,
I still need to test getting so may new collections that the entire
result set is exhausted.
A bug was found with result sets less than the max album size. An album
came back with only two images. When I requested a new collection, the
app crashed with an “array index out of bounds” error, where I tried to
get a Slice out of the result set. Investigating this, now.