Looks like getting new collections is mostly working.
Sometimes it’s hard to tell when images get reused incorrectly (ie.,
the 12-images bug, or something similar) or if the new collections are
entirely unique and correct. But, new images are appearing, and the
cellForItemAtIndexPath() is still going through its logic correctly.
Gonna implement removing cells and come back to inaccuracies with new
collections.