This is almost totally working.
The only problem I’m experiencing now, is that in the milliseconds it
might take to start the task when you drop a pin and when you might tap
it and segue to the photo album VC, the task might not have gotten a
response, yet. If so, the Pin’s photo album is still empty, and the
collection view doesn’t have any items in it.
Also, I need to re-implement the zero results alert.