Deleting the entire photo album will automatically get a new collection when you return to the photo album.
There’s a new bug I found while testing this fix. It appears that if
you visit a photo album, but don’t let all of the cells load images, if
you quit the app and then reload it, when you scroll to a cell that
didn’t have a Core Data image, the app crashes trying to get at the
Photo’s URL property.